refactor(cvc): extract silent-auth + history fetch into shared firestore-history.js
Moves getGoogleAuthTokenSilent/exchangeGoogleToken/silentIdToken/fetchHistory
out of popup.js into firestore-history.js so both the popup and the upcoming
background service worker (in-page SCV highlight) share one implementation.
