+++ clinvar-cvc/popup.html
@@ -177,6 +177,7 @@
   <script src="annotation.js"></script>
   <script src="popup-view.js"></script>
   <script src="history.js"></script>
+  <script src="firestore-history.js"></script>
   <script src="popup.js"></script>
 </body>
 </html>
